Added -o option to wc4 for more gcc-like compilation
This wasn't as easy as it might seem on the surface. This needed calls to
fprintf. However, that uses a FILE*, so for simplicity, I used GNU's
dprintf. However, that led to inefficient behaviour on NFS, so I changed the
Makefile that does the diffing to use a temp file. Then, a bug was uncovered
where the 3rd parameter to open was being ignored, leading stack
misalignment.