Made a start on transforming the IR into SSA ...
A big wart in the current code is that the parser allocates vregs to match what
the backend is doing. For example, the add instruction obliterates the
destination register. To make it worse, a workaround for this was present in
the post ++,-- operator code that had to make a copy of a register. The
approach now is to have pure SSA code and have extra transformation passes to
convert it into something similar to what the backend code generation outputs.
Not only does this lead to more efficient code, it removes the need for the
parser to know about the backend specific quirks. Furthermore, having SSA has
already shown to be useful since a simply optimization was already possible
using this new code: merging of the src2 and dst registers if src2 is no longer
live after the IR statement.