Added constant evaluation and optimized code generation to use contants
Where possible, evaluate constants without generating code. Also, where
possible, generate code where one of the operands is a constant. Fundamentally,
constants are now pushed on the value stack. If anything needs a non-constant,
call pl() or load_constant(). This results in a 7% decrease in generated code
size for a self-compile: 23096 -> 21548 lines.