Don't use registers for a comparison directly followed by a conditional jump
Previous code would always use a register to store the intermediate result. The
new code is much more optimal, but comes at a slight architectural
disadvantage. The backend code assumes that when a conditional jump instruction
that uses CPU flags is encountered, that it is immediately preceded by the
comparison instruction. This is inelegant, but not incorrect, so I'm leaving it
in. This leads to a code reduction of 8% (23450 -> 21485) when self compiling.