Added physical register spilling based on liveness end
When no free physical registers are available, previously the next vreg
would be spilled onto the stack, and the expensive r10/r11 based stack code
would come into play, leading to inefficient code.

First, vregs that has been allocated pregs are checked to see which has the
latest liveness end. Then this is compared with the vreg needing a preg. If the
liveness end of an existing preg is later, then the preg is flushed to the
locals stack. The parser produces an IR in which vregs for locals can be
swapped to be stack based without any rewriting. Effectively falling back to
not using a vreg at all. This leads to a performance increase from 140% to
about 98% compared to gcc. The amount of stack spills should now never be
larger than the version without -O1. There's one drawback though, the redundant
move code has been disabled, since this would otherwise lead to code that the
code generator is unable to handle such as s[0] = s[1] + s[2]. This is going to
be addressed later so that this independent change can be flushed.

Also
- Added print_spilled_register_count
- Added explicit -f options for easier toggling