Fix redundant move optimization by enhancing instruction generation ...
... so that any value can be a local on the stack.

The code is nearly identical to the stack spilling code. I'm not terribly proud
of the duplicated logic, but the functions are sufficiently different for this
to be ok. There's a bit of a plan to merge stack_index with
spilled_stack_index. At that point the functions will be merged.

This fix enabled the redundant move optimization to work again.o