Added linear IR generation by using tree-pattern tile matching
There's way too much going on here to explain in a commit message. The code
produces and IR with x86 instructions. They've not been wired into the backend
code generation yet and there is still a hell of a lot to do. Right now the
only thing that is handled is assignments to local registers and + and *
arithmetic.