Use register allocation for allcate and clobber rax for function calls
This follows the recipe in the book. A function call is assigned to
v1, with a move from v1 to the original dst. v1 is forced into RAX
by adding interference graph edges to all physical registers except RAX.
Live range coalescing is disabled to ensure the mov isn't coalesced in.

A small side effect of this commit is that I discovered that R15 was
never being allocated.

Changes
- Added mov after IR_CALL in IR and removed from codegen
- Removed RAX clobber for all instructions that take place during a function call
- Clobber RAX when dst is not set in IR_CALL
- Force RAX when dst is set in IR_CALL
- Added --debug-graph-coloring for more details on the graph coloring
- Renamed debug_ssa_top_down_register_allocator to debug_register_allocation
- Don't coalesce on the dst vreg used in IR_CALLs
- Fixed total register count 15->16