Split MI4 into MI4 and MPV, making rules more explicit and understandable
MI4 was being used as a signed integer in memory, but also a pointer in memory. The
rules can be tightly split between signed long and pointer operations and so are more
explicit. Besides, it doesn't make sense to have pointer rules such as an indirect
using a signed long.