Fixed binary shift rules + spilling leading to faulty code generation
In the case of an operation instruction, e.g. shrl %cl, %eax, the spilling code may have
to load a spilled dst. Therefore, src2 must be the same as dst.

Furthermore, I removed an unnecessary dst in the initial mov, which was leading to a
useless spill save instruction being added