Added built in offsets to indirect operations ...
The parser now sets offset on pointers in registers when doing struct member lookups.
Quite a heavy change. Although the diff is small, the implications are large. The
number of add instructions in the wcc self compilation have gone from 5190 to 4518. I'm
scared there might be some lurking dragons, but this will do just now.

This paves the way towards being able to implement a memory copy instructions in a
straightforward way as are needed for struct copies.